feat: handle image tag
